set up asset files v7

+++ app/Expense.php
@@ -10,18 +10,18 @@ class Expense extends Model
     protected $primaryKey = 'id';
     protected $table = "expenses";
     protected $cast = [
-        'salarieswages'=>'double',
-        'employeebenefits'=>'double',
-        'allowances'=>'double',
-        'totalps'=>'double',
-        'totalamountmedicine'=>'double',
-        'totalamountmedicalsupplies'=>'double',
-        'totalamountutilities'=>'double',
-        'totalamountnonmedicalservice'=>'double',
-        'totalmooe'=>'double',
-        'amountinfrastructure'=>'double',
-        'amountequipment'=>'double',
-        'totalco'=>'double',
+        'salarieswages'=>'float',
+        'employeebenefits'=>'float',
+        'allowances'=>'float',
+        'totalps'=>'float',
+        'totalamountmedicine'=>'float',
+        'totalamountmedicalsupplies'=>'float',
+        'totalamountutilities'=>'float',
+        'totalamountnonmedicalservice'=>'float',
+        'totalmooe'=>'float',
+        'amountinfrastructure'=>'float',
+        'amountequipment'=>'float',
+        'totalco'=>'float',
         'reportingyear'=>'int'
     ];
 }
